exam result, sutdent photo bug fixes

+++ app/Services/StudentsService.php
@@ -95,7 +95,9 @@ class StudentsService implements StudentsServiceInterface
         // image upload to public/images folder and store image name to database students table
         if ($request->hasFile('imageupload')) {
             $file_path = 'uploads/student/' . $student->image;
-            unlink($file_path);
+            if (file_exists($file_path)) {
+                unlink($file_path);
+            }
             $image = $request->file('imageupload');
             $name = time() . '-' . $request['phone'] . '.' . $image->getClientOriginalExtension();
             $destinationPath = public_path('/uploads/student');
